Postable#post now returns true on success.
Previously the method returned the size of the queue on success but
false on failure. The problem was that on fully parallel systems it was
possible for the method to return 0 on success because the work thread
would remove the message before the #post method returned. Since 0 is
not truthy in Ruby this was misleading. Having #post return true on
success is more reliable and just as informative.