Marked a ThreadLocalVar test pending on Rbx
The test passes on all platforms except Rbx-2 where it consistently
fails. We believe the problem is with the call to GC.start within the
test. As far as we know the code under test *should* work on all
platforms because it is pure Ruby. We cannot say for certain at this
time.