AtomicMarkableReference: compare object identities if not Numeric type
Objects should always be compared by reference so we can ensure that we are
referring to the *same* object, not simply one that maybe be considered
logically equivalent by overriding `==`.

This is, of course, unless it is a Numeric. When comparing something primitive
like a number, we expect that the equality operator makes mathematical sense.