Improved java extension loading.
The original implementation assumed that the Java extensions would
always be compiled and loaded when running on JRuby. This is not always
the case. During development and testing if the gem in included directly
from source, as when the Gemfile references git directly, none of the
native extensions will exist. In production if the pure-Ruby gem
is loaded under JRuby the extensions won't exist, either.

This commit updates Java extension loading to follow the existing
pattern for C extensions. Relevant parts of the code check to see if the
extensions have been correctly loaded first, rather than assuming they
are. The test suite still assumes that Java and C extensions are loaded,
however. This is necessary to catch bugs which prevent extensions from
properly compiling and loading.