Add ready-length attr_reader
The ready queue is available via a local instance variable. Exposing
the length of the array of ready workers will allow the health of the
pool to be measured/monitored as tasks are moving through the pool.

Important values are total workers (`length`), idle workers
(`ready_length`), and busy workers (difference between the 2).