Fixes #641 - remove sharp dependency edge -> core
Since we follow semantic version of concurrent-ruby, the changes
in '1.0.x' should not influence the current edge version and we
should not have have the sharp dependency there. This makes it easier
for people brave enough to give the edge a try to keep up with
the changes while keeping their projects working with others that are
on the same core version.