Fix Map#each and #each_pair not returning enumerator outside of MRI
When no block was passed to Map#each_pair or its alias Map#each, the
NonConcurrentMapBackend would return an enumerator, which would allow
enumerable methods to be used on a Map.

But the alternate Map backends, JRubyMapBackend and
AtomicReferenceMapBackend did not implement the same logic.

As a fix, let's move the logic down to the shared Map class, so all
implementations can use it.

Fixes #643