Workaround struct inspect bug on JRuby 9.2.0.0
Due to a bug in JRuby 9.2.0.0, mutating the result of `Module#to_s`
actually changes the result of other `Module#to_s` calls to the same
module or class, rather than returning a new `String`.

This broke `AbstractStruct#pr_underscore` which is used when rendering
`#inspect` for structures.

This was already fixed upstream, but as a workaround, let's dup the
String before we mutate it, so as to not trip anyone using this JRuby
version (and to restore Travis to beautiful green).

Issue https://github.com/jruby/jruby/issues/5229