TimerTask: kill the execution when it exceeds the task timeout and add timer_task timeout specs
Before this commit when a TimerTask task exceeded the timeout the job
kept running. That could lead to thread leaking. Related to #639

This PR changes the TimerTask executor to be a SingleThreadExecutor. So
whenever the task doesn't finish in time, the execution thread is killed
(in this case the pool is killed)

The spec ensures that this new behavior is working correctly. It fails
on master. If the main job isn't killed after the timeout, the
latch.wait(2) would raise, since the main task sleeps for 5 seconds.

Closes #639