make the pruning of thread pools a public API
This makes the pruning of thread pools a public API, which may
or may not be a no-op, depending on the underlying thread pool
implementation.

It also reworks the Ruby implementation to be more aggressive in
the pruning - it will kill all to-be-pruned threads in one call
instead of doing them one by one. Moreover, workers do not get
woken up to check for their idleness anymore, the thread of the
caller does it instead.