Use Ruby's Etc.nprocessors if available
MRI has had Etc.nprocessors since Ruby 2.2, which uses:
- sched_getaffinity(): Linux
- sysconf(_SC_NPROCESSORS_ONLN): GNU/Linux, NetBSD, FreeBSD, OpenBSD, DragonFly BSD, OpenIndiana, Mac OS X, AIX
- GetSystemInfo(): Win32

This should be faster than the alternatives of reading from
/proc/cpuinfo, using WIN32OLE, or spawning a new process.

This also has the advantage of reporting processor count based on CPU
affinity which /usr/bin/nproc did, but reading /proc/cpuinfo does not.
When using this value for controlling the level of parallelism, we
should ideally be using the processes CPU affinity.